+++ feature/settings/src/main/java/com/melhoreapp/feature/settings/ui/SettingsViewModel.kt
@@ -91,7 +91,7 @@ class SettingsViewModel @Inject constructor(
                 toDelete.forEach { reminder ->
                     reminderScheduler.cancelReminder(reminder.id)
                     reminderDao.deleteById(reminder.id)
-                    syncRepository.deleteReminderFromCloud(uid, reminder.id)
+                    if (uid != "local") syncRepository.deleteReminderFromCloud(uid, reminder.id)
                 }
             }
         }
